Fix padding for AU pseudo IBANs
When we calculate pseudo-ibans, if the account number is shorter than
the expected lenght, the pseudo-iban is right-justified with `X`. The
leading `X`s are then stripped on the splitter.

That worked well for Sweedish accounts, since the account number is only
digits.

However, in Australia, account numbers are alphanumeric, which means an
account number can start with `X`.

We had to use a different padding for Australia but keep using `X` for
backwards compatibility of Sweedish pseudo ibans.