Update iban_spec.rb for CA
https://gocardless.myjetbrains.com/youtrack/issue/BANKINT-147

leaving IBAN validation check for later,
till we actually need it.

ibandit account number length validation
can check for maximum length but not
against a range, say, 7 to 12 digits,
allowed in CA.

given we are only going to deal with
pseudo-IBANs in CA, choosing to not
complete IBAN validation unless
we need it.