Update local_details_cleaner for CA
https://gocardless.myjetbrains.com/youtrack/issue/BANKINT-147

so that we generate valid IBANs
for CA that allow for a minimum
account number length of 7, and
a maximum of 12.

we do this by validating that the
length is 12, and padding account
number with 0s till it reaches
that length.

a workaround for not having
account number validation that
checks for length in a range.