Bump gc_ruboconfig to 2.5.0
v2.4.0 of gc_ruboconfig had the Enabled property of the Style/Encoding property set to `never`. As Rubocop now verifies that this property must be a boolean, the following error is encountered when attempting to run Rubocop >= 0.75 on this project.

```
The Property Enabled of cop Style/Encoding is supposed to be a boolean and never is not
```

This has been fixed in gc_ruboconfig v2.5.0 (see https://github.com/gocardless/gc_ruboconfig/pull/55/), thus necessitating this bump