Tighten up validation for Canadian account numbers
Previously this validation would allow account number formats like

- `1234567-789`
- `1234567-ABC`

This is because calling `RegEx.match?` will return the first match it
sees rather than testing the entire string. So to fix this we use the
start and end of string matches to ensure the entire string is tested.

There are likely other formats affected by the same issue, but as we
haven't seen these in the wild I'm holding off fixing these.